Add files via upload
NLTK sentiment analysis, featuring vader. Now with Java!
RunsPython creates a process that does a command line call to the python file. The command line argument is a pathname to a file containing the emails as strings, one per line. From there, python returns an average of the sentiment, currently. The current speed for 1000 emails of character length 3009 is ~10.5s.