Add files via upload
This should be the final version for these. See comments for specific usage instruction. If that's too much for you, it goes like this: RunsPython has a static method that takes a string array of all the emails you want it to analyze. It returns an array of doubles, size 4, where 0:neg, 1:neu, 2:pos, 3:compound. 
Included is an example of the encrypted text file. We can change the final keyword and shift value for launch, or even change up the method so that it takes a keyword and number for the ciphers. Either is fine by me.